fix: 帧计时 m_LastFrameTime 从 float 改为 double
glfwGetTime() 返回 double，截断为 float 后在长时间运行
（~4.6 小时后 float 精度仅 ~0.5ms）导致 delta 计算退化。
改为 double 存储累积时间，delta 计算完后再转 float 给 Timestep。

+++ Engine/src/Core/Application.cpp
@@ -130,11 +130,11 @@ namespace Engine
     {
         while (m_Running)
         {
-            float time = static_cast<float>(glfwGetTime());
-            Timestep timestep = time - m_LastFrameTime;
+            double time = glfwGetTime();
+            Timestep timestep = static_cast<float>(time - m_LastFrameTime);
             m_LastFrameTime = time;
 
-            PerformanceMonitor::Get().BeginFrame(time);
+            PerformanceMonitor::Get().BeginFrame(static_cast<float>(time));
 
             if (!m_Minimized)
             {
